+++ odorik/test_main.py
@@ -183,6 +183,13 @@ class TestCommands(TestCase):
         self.assertTrue('price: 0.1484' in output)
 
     @httpretty.activate
+    def test_data_all(self):
+        """Test getting data summary for number"""
+        register_uris()
+        output = execute(['mobile-data', '--all'])
+        self.assertTrue('price: 0.1484' in output)
+
+    @httpretty.activate
     def test_data_list(self):
         """Test getting data list"""
         register_uris()